fix refactor pacote domain e exclusao de migration pacote infra
