+++ api/src/main/java/med/voll/api/domain/medico/DadosAtualizacaoMedico.java
@@ -1,7 +1,7 @@
-package med.voll.api.medico;
+package med.voll.api.domain.medico;
 
 import jakarta.validation.constraints.NotNull;
-import med.voll.api.endereco.Endereco;
+import med.voll.api.domain.endereco.Endereco;
 
 public record DadosAtualizacaoMedico(
 		@NotNull
