+++ api/src/main/java/med/voll/api/domain/medico/DadosListagemMedico.java
@@ -1,4 +1,4 @@
-package med.voll.api.medico;
+package med.voll.api.domain.medico;
 
 public record DadosListagemMedico(Long id,String nome, String email, String crm, Especialidade especialidade) {
 public DadosListagemMedico(Medico medico) {
