+++ api/src/main/java/med/voll/api/domain/medico/DadosdetalhadoMedico.java
@@ -1,6 +1,6 @@
-package med.voll.api.medico;
+package med.voll.api.domain.medico;
 
-import med.voll.api.endereco.Endereco;
+import med.voll.api.domain.endereco.Endereco;
 
 public record DadosdetalhadoMedico(Long id,
 String nome,
