+++ api/src/main/java/med/voll/api/domain/paciente/DadosDetalhamentoPaciente.java
@@ -1,6 +1,6 @@
-package med.voll.api.paciente;
+package med.voll.api.domain.paciente;
 
-import med.voll.api.endereco.Endereco;
+import med.voll.api.domain.endereco.Endereco;
 
 public record DadosDetalhamentoPaciente(String nome, String email, String telefone, String cpf, Endereco endereco) { 
     public DadosDetalhamentoPaciente(Paciente paciente) {
