--- api/src/main/java/med/voll/api/endereco/DadosAtualizacaoEndereco.java
@@ -1,5 +1,0 @@
-package med.voll.api.endereco;
-
-public class DadosAtualizacaoEndereco {
-
-}
