+++ api/src/main/java/med/voll/api/infra/TratadorDeErros.java
@@ -1,0 +1,14 @@
+package med.voll.api.infra;
+
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import jakarta.persistence.EntityNotFoundException;
+
+@RestControllerAdvice
+public class TratadorDeErros {
+	@ExceptionHandler(EntityNotFoundException.class)
+public void tratarError404() {
+	
+}
+}
