--- api/src/main/resources/db/migration/V2__alter-table-madicos-add-column-telefone.sql
@@ -1,1 +1,0 @@
-alter table medicos add telefone varchar(20) not null;
