--- api/src/main/resources/db/migration/V3__create-table-pacientes.sql
@@ -1,16 +1,0 @@
-create table pacientes(
-    id bigint not null auto_increment,
-    nome varchar(100) not null,
-    email varchar(100) not null unique,
-    cpf varchar(14) not null unique,
-    telefone varchar(20) not null,
-    logradouro varchar(100) not null,
-    bairro varchar(100) not null,
-    cep varchar(9) not null,
-    complemento varchar(100),
-    numero varchar(20),
-    uf char(2) not null,
-    cidade varchar(100) not null,
-
-    primary key(id)
-);
