--- api/src/main/resources/db/migration/V5__alter-table-pacientes-add-ativo.sql
@@ -1,3 +1,0 @@
-alter table pacientes add column ativo tinyint;
-update pacientes set ativo = 1;
-alter table pacientes modify ativo tinyint not null;
